Improve /assets/?q= to allow searching asssets by both tags and description
In the beginning with DB queries we allowed filtering by tags, after switching to ES we started doing full text search
on description instead of using tags, but it helps to use both description and tags for searching. Tags should serve
as a hint to the user and should function as a keyword search instrument along with full text search on description.