Disable auth requirement for connecting social accounts (#208)
Why: We trust LinkedIn, Google to validate the user email, we don't want to have the user to login first to be able to connect with their social media account. Caveat: We want to have email verifications for regular accounts so that lack of auth in connect is not misused.