Add data models related to solutions and solution bookings.
Note: We already had the data model, this changed the relation between Asset and Solution to an M2M,
since it is not necessary that a Solution will be tied to an asset. A solution can span multiple assets and one asset can have multiple related solutions. Also, added SolutionBooking model which represents one instnace of a solution being provided to a specific user.