Create solution if it doesn't already exist, durign product updates
While this is a rare case, it may happen if a product already exists (pre webhook handler logic addition), and for cases where create events weren't handled before,
due to the server temporarily being down or some other reason. In these cases updating a product will trigger a Solution creation if it does not exist.